fix: move the shared components FilterList and the injection option, into the shared components dir
